Correct format to display distance

+++ RentTool/Models/CardTool.cs
@@ -7,6 +7,6 @@ namespace RentTool.Models
         public string name { get; set; }
         public string image { get; set; }
         public string pricePerDay { get; set; }
-        public float distance { get; set; }
+        public double distance { get; set; }
     }
 }
